+++ docs/ComponentGPT.md
@@ -1,0 +1,220 @@
+Хорошо, тогда давай я перепишу твою доку в вылизанный, читабельный и единообразный вариант — с нормальной терминологией, структурой и исправленными опечатками.
+
+---
+
+# Component System Specification
+
+## 1. Общая концепция
+
+**Component** — базовый строительный блок TUI-приложения.
+Каждый компонент — это отдельная структура, способная:
+
+* хранить собственное состояние,
+* обрабатывать события,
+* генерировать сообщения (Messages),
+* отрисовывать себя и своих дочерних компонентов.
+
+Компоненты образуют **иерархическое дерево**: один компонент может содержать несколько дочерних.
+Обмен данными между компонентами происходит через **Messages**.
+
+---
+
+## 2. Жизненный цикл компонента
+
+### 2.1. Сообщения (Messages)
+
+Каждый компонент определяет собственный `enum` с суффиксом `Msg`:
+
+```rust
+enum MyComponentMsg {
+    Submitted(u16),
+    Canceled,
+}
+```
+
+**Правила:**
+
+* Имя формируется как `<ИмяКомпонента>Msg`.
+* Сообщения должны описывать завершённые действия компонента.
+* Возвращаемый тип обработки событий — `Option<MyComponentMsg>`.
+
+---
+
+### 2.2. Поля (Fields)
+
+Компонент хранит любое количество полей для своей логики:
+
+```rust
+count: u8,
+title: String,
+buffer: String,
+manager: ContactsManager,
+```
+
+**Особое поле**:
+Для компонентов, которые могут получать фокус, обязательно наличие:
+
+```rust
+focused: bool
+```
+
+---
+
+### 2.3. Дочерние компоненты (Child Components)
+
+* Компонент может содержать один или несколько подкомпонентов.
+* В случае нескольких **focusable**-компонентов активным может быть только один.
+* Родитель может отрисовывать как дочерние компоненты, так и собственные элементы вокруг них.
+
+---
+
+## 3. Основные методы
+
+### 3.1. Конструкторы и утилиты
+
+* `new()` — для создания компонента.
+* Геттеры/сеттеры для внутренних полей.
+* Методы для манипуляции состоянием.
+
+---
+
+### 3.2. Отрисовка (`draw`)
+
+Каждый компонент обязан иметь метод:
+
+```rust
+fn draw(&mut self, frame: &mut Frame, area: Rect);
+```
+
+**Параметры:**
+
+* `frame` — объект, предоставляющий доступ к буферу отрисовки (`frame.buffer_mut()`).
+* `area` — область терминала, отведённая под компонент.
+
+**Особенности:**
+
+* Компонент может скрывать/показывать дочерние элементы.
+* Поведение может зависеть от `focused`.
+
+---
+
+### 3.3. Обработка событий (Event Routing)
+
+#### Алгоритм:
+
+1. Получить `Event` от родителя.
+2. Выполнить **component-level** обработку.
+3. Если возвращено `Some(Message)` — прекратить дальнейшую обработку.
+4. Если нет — передать событие активному дочернему компоненту.
+5. Обернуть его сообщение в собственный `Msg` и вернуть родителю.
+
+**Пример передачи события дочернему компоненту:**
+
+```rust
+self.sub_component
+    .handle_event(event)
+    .map(ComponentMsg::SubComponent)
+```
+
+---
+
+### 3.4. Обработчики событий (Event Handlers)
+
+Методы, которые реагируют на конкретные действия, например:
+
+```rust
+fn increment(&mut self) -> Option<MyComponentMsg> {
+    self.count += 1;
+    Some(MyComponentMsg::Incremented)
+}
+```
+
+**Особенности:**
+
+* Не принимают параметров (кроме `&mut self`).
+* Возвращают `Option<Msg>`.
+
+---
+
+### 3.5. Обработчики сообщений (Message Handlers)
+
+**(В разработке)**
+Используются в основном в корневом компоненте (`App`), куда поднимаются все сообщения от дочерних элементов.
+
+---
+
+## 4. Реализация стандартных трейтов
+
+Для удобства можно реализовывать:
+
+* `Default`
+* `From<T>`
+* и другие, упрощающие создание и преобразование компонентов.
+
+---
+
+## 5. Пример минимального компонента
+
+```rust
+enum CounterMsg {
+    Incremented,
+    Decremented,
+}
+
+struct Counter {
+    focused: bool,
+    value: i32,
+}
+
+impl Counter {
+    fn new() -> Self {
+        Self { focused: false, value: 0 }
+    }
+
+    fn draw(&mut self, frame: &mut Frame, area: Rect) {
+        let text = format!("Counter: {}", self.value);
+        frame.render_widget(text, area);
+    }
+
+    fn handle_event(&mut self, event: Event) -> Option<CounterMsg> {
+        match event {
+            Event::Key(KeyEvent { code: KeyCode::Up, .. }) => {
+                self.value += 1;
+                Some(CounterMsg::Incremented)
+            }
+            Event::Key(KeyEvent { code: KeyCode::Down, .. }) => {
+                self.value -= 1;
+                Some(CounterMsg::Decremented)
+            }
+            _ => None,
+        }
+    }
+}
+```
+
+---
+
+## 6. Рекомендации по именованию
+
+**Хорошо:**
+
+* `TaskForm`
+* `TextField`
+* `WeatherInfo`
+
+**Плохо:**
+
+* `TransactionToOutcomeWidget`
+* `PersonInformationCompoment`
+
+---
+
+## 7. Идеи и расширения
+
+* Макросы для генерации трейтов компонентов.
+* Автоматическая маршрутизация сообщений.
+* Глобальный менеджер фокуса.
+
+---
+
+Если хочешь, я могу к этому документу ещё добавить **схему потока событий** (как Event идёт от App до компонента и обратно в виде Message), чтобы стало совсем наглядно. Думаю, она бы отлично вписалась в эту спецификацию.
