+++ src/tui/mod.rs
@@ -1,0 +1,5 @@
+mod widgets;
+mod app;
+mod run;
+
+pub use run::run_tui;
