+++ src/tui/widgets/mod.rs
@@ -2,3 +2,5 @@ pub mod component;
 pub mod input_field;
 pub mod task_info;
 pub mod tasks_list;
+pub mod task_form;
+pub mod button;
